pyproject.toml
[project]
name = "MusicTerm"
version = "0.1.1"
authors = [{ name = "ZachVFX", email = "zachallrd@gmail.com" }]
description = "play music from your termninal."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "just-playback>=0.1.8",
    "msgspec>=0.19.0",
    "textual>=6.1.0",
    "tomli-w>=1.2.0",
    "ytmusicapi>=1.10.1",
    "textual-image[textual]>=0.8.4",
    "pillow>=11.1.0",
    "requests-cache>=1.2.1",
    "winrt-windows-foundation>=3.2.1; sys_platform == 'win32'",
    "winrt-windows-media-playback>=3.2.1; sys_platform == 'win32'",
    "winrt-windows-data-xml-dom>=3.2.1; sys_platform == 'win32'",
    "winrt-windows-ui-notifications>=3.2.1; sys_platform == 'win32'",
    "winrt-windows-media-core>=3.2.1; sys_platform == 'win32'",
    "winrt-windows-media>=3.2.1; sys_platform == 'win32'",
    "winrt-windows-storage>=3.2.1; sys_platform == 'win32'",
    "winrt-windows-storage-streams>=3.2.1; sys_platform == 'win32'",
    "winrt-windows-foundation-collections>=3.2.1; sys_platform == 'win32'",
    "music-tag>=0.4.3",
    "pypresence",
    "lrcup>=0.10.1",
    "dbus-fast>=1.83.0",
    "yt-dlp>=2025.12.08",
    "textual-serve>=1.1.2",
    "certifi>=2025.8.3",
    "python-dotenv>=1.2.1",
]

[tool.pytest.ini_options]
pythonpath = [
  "src",
]

[tool.uv.sources]
pypresence = { url = "https://github.com/qwertyquerty/pypresence/archive/master.zip" }

[project.urls]
Homepage = "https://github.com/ZachVFXX/PyMusicTerm"
Issues = "https://github.com/ZachVFXX/PyMusicTerm/issues"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "textual-dev>=1.7.0",
]
